% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runif_r}
\alias{runif_r}
\title{Random Uniform Number Generator using base R}
\usage{
runif_r(n, min = 0, max = 1)
}
\arguments{
\item{n}{An \code{unsigned integer} denoting the number of realizations to generate.}

\item{min}{A \code{double} indicating the minimum \eqn{a} value 
in the uniform's interval \eqn{\left[a,b\right]}}

\item{max}{A \code{double} indicating the maximum \eqn{b} value 
in the uniform's interval \eqn{\left[a,b\right]}}
}
\description{
The function provides an alternative implementation of random uniform distribution
sampling using R's rng scope.
}
\examples{
set.seed(134)
b = runif_r(10)
}
