% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitmo-package.R
\docType{package}
\name{sitmo-package}
\alias{sitmo}
\alias{sitmo-package}
\title{sitmo: Parallel Pseudo Random Number Generator (PPRNG) 'sitmo' Header Files}
\description{
Provided within is a high quality and fast PPRNG that is able to be used in an 'OpenMP' parallel
environment compiled under either C++98 or C++11. The objective of this package release is to consolidate
the distribution of the 'sitmo' library on CRAN by enabling others to link to the 'sitmo' header file instead
of including a copy of 'sitmo' within their individual package. Lastly, the package contains example
implementations using 'sitmo' and three accompanying vignette that provide additional information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/sitmo}
  \item \url{https://www.sitmo.com/?p=1206}
  \item Report bugs at \url{https://github.com/coatless/sitmo/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item Thijs van den Berg \email{thijs@sitmo.com} [copyright holder]
}

}
