% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sitmo_draws}
\alias{sitmo_draws}
\title{Example RNG Draws with sitmo}
\usage{
sitmo_draws(n)
}
\arguments{
\item{n}{A \code{unsigned int} is a .}
}
\value{
A \code{vec} with random sequences.
}
\description{
Shows a basic setup and use case for sitmo.
}
\examples{
n = 10
a = sitmo_draws(n)
}
