% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sitmo_engine_reset}
\alias{sitmo_engine_reset}
\title{Example Seed Set and RNG Draws with sitmo}
\usage{
sitmo_engine_reset(n, seed)
}
\arguments{
\item{n}{An \code{unsigned int} that dictates how many realizations occur.}

\item{seed}{An \code{unsigned int} that controls the rng seed.}
}
\value{
A \code{matrix} with random sequences.
}
\description{
Shows how to set a seed in sitmo.
}
\examples{
n = 10
a = sitmo_engine_reset(n, 1337)

isTRUE(all.equal(a[,1],a[,2]))
}
