% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sitmo_engine_seed}
\alias{sitmo_engine_seed}
\title{Example Seed Set and RNG Draws with sitmo}
\usage{
sitmo_engine_seed(n, seed)
}
\arguments{
\item{n}{An \code{unsigned int} that dictates how many realizations occur.}

\item{seed}{An \code{unsigned int} that controls the rng seed.}
}
\value{
A \code{vector} with random sequences.
}
\description{
Shows how to set a seed in sitmo.
}
\examples{
n = 10
a = sitmo_engine_seed(n, 1337)
b = sitmo_engine_seed(n, 1337)
c = sitmo_engine_seed(n, 1338)

isTRUE(all.equal(a,b))
isTRUE(all.equal(a,c))
}
