% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sitmo_example}
\alias{sitmo_example}
\title{Test Generation of a Random Exponential using sitmo and C++11}
\usage{
sitmo_example(n, lambda, seeds)
}
\arguments{
\item{n}{An \code{unsigned integer} denoting the number of realizations to generate.}

\item{lambda}{A \code{double} indicating the scale parameter of the distribution.}

\item{seeds}{A \code{vec} containing a list of seeds. Each seed is run on its own core.}
}
\value{
A \code{vec} containing the realizations.
}
\description{
The function provides an implementation of creating realizations
an exponential
}
\details{
The following function's true power is only accessible on platforms that support OpenMP (e.g. Windows and Linux).
However, it does provide a very good example as to how to make ones code applicable across multiple platforms.

With this being said, how we determine how many cores to split the generation to is governed by the number of seeds supplied.
In the event that one is using OS X, only the first seed supplied is used.
}
\examples{
a = sitmo_example(10, 5.0, c(1))

b = sitmo_example(10, 5.0, c(1))

c = sitmo_example(10, 5.0, c(2))

all.equal(a,b)

all.equal(a,c)
}

